+++ src/apis/upload/index.ts
@@ -1,1 +1,2 @@
 export { uploadImage } from './image-upload';
+export { uploadImages } from './images-upload';
